fixed calculation of selection box intersection with selectable elements to work with translated and rotated helix groups
