+++ lib/src/reducers/selection_reducer.dart
@@ -1,20 +1,16 @@
-import 'dart:html';
-import 'dart:svg';
-
 import 'package:redux/redux.dart';
 import 'package:built_collection/built_collection.dart';
 import '../state/edit_mode.dart';
 import '../state/select_mode.dart';
 
 import '../state/selectable.dart';
-import '../view/design.dart';
 import '../state/helix.dart';
 import '../state/app_state.dart';
 import '../state/selection_box.dart';
 import '../actions/actions.dart' as actions;
 import '../util.dart' as util;
+import '../middleware/selections_intersect_box_compute.dart' as select;
 import '../constants.dart' as constants;
-import '../state/selectable.dart';
 import 'util_reducer.dart';
 
 SelectablesStore selectables_store_reducer(SelectablesStore selectables_store, AppState state, action) {
@@ -28,7 +24,7 @@ SelectablesStore selectables_store_reducer(SelectablesStore selectables_store, A
 
 GlobalReducer<SelectablesStore, AppState> selectables_store_global_reducer = combineGlobalReducers([
   TypedGlobalReducer<SelectablesStore, AppState, actions.Select>(select_reducer),
-  TypedGlobalReducer<SelectablesStore, AppState, actions.SelectionsAdjust>(selections_adjust_reducer),
+  TypedGlobalReducer<SelectablesStore, AppState, actions.SelectOrToggleItems>(select_or_toggle_items_reducer),
   TypedGlobalReducer<SelectablesStore, AppState, actions.SelectAllSelectable>(select_all_selectables_reducer),
 ]);
 
@@ -97,46 +93,9 @@ SelectablesStore select_all_selectables_reducer(
   return selectables_store.select_all(selected);
 }
 
-SelectablesStore selections_adjust_reducer(
-    SelectablesStore selectables_store, AppState state, actions.SelectionsAdjust action) {
-  //XXX: this is not a pure reducer since it looks at SVG view properties to determine SVG object
-  // intersection.
-  // It seems difficult to put enough state into AppState to reliably detect when one SVG element
-  // overlaps another in a way that is robust to future changes in how we draw.
-  // Maybe figure out some way to do this with middleware so the reducer can remain pure.
-
-  RectElement select_box = querySelector('#selection-box-main') as RectElement;
-  if (select_box == null) {
-    return selectables_store;
-  }
-  Rect select_box_bbox = select_box.getBBox();
-
-  //XXX: Firefox does not have getIntersectionList or getEnclosureList
-  // (no progress for 10 years on that: https://bugzilla.mozilla.org/show_bug.cgi?id=501421)
-  // Besides, it didn't work well in Chrome and I basically had to implement it myself based on bounding boxes.
-
-  bool is_origami = state.design.is_origami;
-  var select_modes = state.ui_state.select_mode_state.modes;
-  Set<SvgElement> elts_overlapping =
-      util.enclosure_list_in_elt(MAIN_VIEW_SVG_ID, select_box_bbox, select_modes, is_origami).toSet();
-
-  var selectable_by_id = state.design.selectable_by_id;
-  List<Selectable> overlapping_now = [
-//    for (var elt in elts_overlapping) if (selectables_by_id.containsKey(elt.id)) selectables_by_id[elt.id]
-    for (var elt in elts_overlapping) if (selectable_by_id.containsKey(elt.id)) selectable_by_id[elt.id]
-  ];
-
-  List<Selectable> overlapping_now_select_mode_enabled = [];
-  for (var obj in overlapping_now) {
-    if (state.ui_state.select_mode_state.is_selectable(obj)) {
-      overlapping_now_select_mode_enabled.add(obj);
-    }
-  }
-
-  return action.toggle
-      ? selectables_store.toggle_all(overlapping_now_select_mode_enabled)
-      : selectables_store.select_all(overlapping_now_select_mode_enabled);
-}
+SelectablesStore select_or_toggle_items_reducer(
+        SelectablesStore selectables_store, AppState state, actions.SelectOrToggleItems action) =>
+    action.toggle ? selectables_store.toggle_all(action.items) : selectables_store.select_all(action.items);
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 // selectables local reducer
@@ -168,14 +127,15 @@ GlobalReducer<BuiltSet<int>, AppState> side_selected_helices_global_reducer = co
 
 BuiltSet<int> helix_selections_adjust_reducer(
     BuiltSet<int> helix_idxs_selected, AppState state, actions.HelixSelectionsAdjust action) {
+  //FIXME: this reducer isn't pure. Move into middleware similar to selections_intersect_box_compute
   bool toggle = action.toggle;
   var selection_box = action.selection_box;
   var all_helices_in_displayed_group = state.design.helices_in_group(state.ui_state.displayed_group_name);
-  List<util.Box> all_bboxes =
+  List<select.Box> all_bboxes =
       all_helices_in_displayed_group.values.map((helix) => helix_to_box(helix)).toList();
-  var selection_box_as_box = util.Box.from_selection_box(selection_box);
+  var selection_box_as_box = select.Box.from_selection_box(selection_box);
   List<Helix> helices_overlapping =
-      util.enclosure_list(all_helices_in_displayed_group.values, all_bboxes, selection_box_as_box);
+    select.enclosure_list(all_helices_in_displayed_group.values, all_bboxes, selection_box_as_box);
 //      util.intersection_list(all_helices.toList(), all_bboxes, util.Box.from_selection_box(selection_box));
   List<int> helix_idxs_overlapping = helices_overlapping.map((helix) => helix.idx).toList();
 
@@ -197,7 +157,7 @@ BuiltSet<int> helix_selections_adjust_reducer(
   return helices_idxs_selected_new.build();
 }
 
-util.Box helix_to_box(Helix helix) {
+select.Box helix_to_box(Helix helix) {
   //FIXME: this is making boxes that are not far enough apart
   var position3d = helix.position3d();
   num x, y, width, height;
@@ -205,7 +165,7 @@ util.Box helix_to_box(Helix helix) {
   x = svg_pos.x - constants.HELIX_RADIUS_SIDE_PIXELS;
   y = svg_pos.y - constants.HELIX_RADIUS_SIDE_PIXELS;
   height = width = constants.HELIX_RADIUS_SIDE_PIXELS * 2.0;
-  return util.Box(x, y, width: width, height: height);
+  return select.Box(x, y, width: width, height: height);
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
