+++ lib/src/util.dart
@@ -1209,169 +1209,6 @@ String wc_base(String base) {
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// intersection geometry
-
-class Box {
-  num height;
-  num width;
-  num x;
-  num y;
-
-  factory Box.from(Rect svg_rect) =>
-      Box(svg_rect.x, svg_rect.y, width: svg_rect.width, height: svg_rect.height);
-
-  factory Box.from_selection_box(SelectionBox box) => Box(box.x, box.y, width: box.width, height: box.height);
-
-  Box(this.x, this.y, {num height = null, num width = null, num x2 = null, num y2 = null}) {
-    if (width == null && x2 == null) {
-      throw ArgumentError('at least one of height or x2 must be non-null');
-    } else if (x2 == null) {
-      this.width = width;
-    } else if (width == null) {
-      this.width = x2 - x;
-    }
-
-    if (height == null && y2 == null) {
-      throw ArgumentError('at least one of height or x2 must be non-null');
-    } else if (y2 == null) {
-      this.height = height;
-    } else if (height == null) {
-      this.height = y2 - y;
-    }
-  }
-
-  num get x2 => x + width;
-
-  num get y2 => y + height;
-
-  set x2(num x2new) {
-    width = x2new - x;
-  }
-
-  set y2(num y2new) {
-    height = y2new - y;
-  }
-}
-
-// gets list of elements associated to Selectables that intersect select_box_bbox
-List<E> intersection_list<E>(List<E> elts, List<Box> bboxes, Box select_box) =>
-    generalized_intersection_list(elts, bboxes, select_box, intervals_overlap);
-
-// gets list of elements associated to Selectables that intersect select_box_bbox
-List<E> enclosure_list<E>(Iterable<E> elts, List<Box> bboxes, Box select_box) =>
-    generalized_intersection_list(elts, bboxes, select_box, interval_contained);
-
-// indicates if (l1,h1) intersect (l2,h2) \neq empty
-bool intervals_overlap(num l1, num h1, num l2, num h2) {
-  return h1 >= l2 && h2 >= l1;
-}
-
-// indicates if (l1,h1) \subseteq (l2,h2)
-bool interval_contained(num l1, num h1, num l2, num h2) {
-  return l1 >= l2 && h1 <= h2;
-}
-
-List<E> generalized_intersection_list<E>(
-    Iterable<E> elts, List<Box> bboxes, Box select_box, bool overlap(num l1, num h1, num l2, num h2)) {
-  if (elts.length != bboxes.length) {
-    throw ArgumentError(
-        'elts (length ${elts.length}) and bboxes (length ${bboxes.length}) must have same length');
-  }
-  List<E> elts_intersecting = [];
-//  for (int i = 0; i < elts.length; i++) {
-  int i = 0;
-  for (E elt in elts) {
-    Box elt_bbox = bboxes[i++];
-    if (boxes_intersect_generalized(elt_bbox, select_box, overlap)) {
-      elts_intersecting.add(elt);
-    }
-  }
-  return elts_intersecting;
-}
-
-bool boxes_intersect_generalized(Box elt_bbox, Box select_box, bool overlap(num l1, num h1, num l2, num h2)) {
-  num elt_x2 = elt_bbox.x + elt_bbox.width;
-  num select_box_x2 = select_box.x + select_box.width;
-  num elt_y2 = elt_bbox.y + elt_bbox.height;
-  num select_box_y2 = select_box.y + select_box.height;
-  return overlap(elt_bbox.x, elt_x2, select_box.x, select_box_x2) &&
-      overlap(elt_bbox.y, elt_y2, select_box.y, select_box_y2);
-}
-
-//// gets list of elements associated to Selectables that intersect select_box_bbox in elements with classname
-//List<SvgElement> intersection_list_in_elt(String classname, Rect select_box_bbox) {
-//  return generalized_intersection_list_in_elt(classname, select_box_bbox, intervals_overlap);
-//}
-
-// gets list of elements associated to Selectables that intersect select_box_bbox in elements with classname
-List<SvgElement> enclosure_list_in_elt(
-    String classname, Rect select_box_bbox, Iterable<SelectModeChoice> select_modes, bool is_origami) {
-  return generalized_intersection_list_in_elt(
-      classname, select_box_bbox, select_modes, is_origami, interval_contained);
-}
-
-generalized_intersection_list_in_elt(String classname, Rect select_box_bbox,
-    Iterable<SelectModeChoice> select_modes, bool is_origami, bool overlap(num l1, num h1, num l2, num h2)) {
-  List<SvgElement> elts_intersecting = [];
-
-//  List<Element> selectable_elts = querySelectorAll('.selectable');
-  List<Element> selectable_elts = find_selectable_elements(select_modes, is_origami);
-
-  for (GraphicsElement elt in selectable_elts) {
-    Rect elt_bbox = elt.getBBox();
-//    util.transform_rect_svg_to_mouse_coord_main_view(elt_bbox);
-    if (bboxes_intersect_generalized(elt_bbox, select_box_bbox, overlap)) {
-      elts_intersecting.add(elt);
-    }
-  }
-  return elts_intersecting;
-}
-
-List<Element> find_selectable_elements(Iterable<SelectModeChoice> select_modes, bool is_origami) {
-  List<SelectModeChoice> select_modes_not_scaffold_or_staple = [
-    for (var mode in select_modes)
-      if (mode != SelectModeChoice.scaffold && mode != SelectModeChoice.staple) mode
-  ];
-  if (select_modes_not_scaffold_or_staple.isEmpty) {
-    return [];
-  }
-
-  List<String> selectors = [];
-  if (is_origami &&
-      (!select_modes.contains(SelectModeChoice.scaffold) ||
-          !select_modes.contains(SelectModeChoice.staple))) {
-    if (select_modes.contains(SelectModeChoice.scaffold)) {
-      for (var mode in select_modes_not_scaffold_or_staple) {
-        selectors.add('.${SelectModeChoice.scaffold.css_selector()}.${mode.css_selector()}');
-      }
-    } else if (select_modes.contains(SelectModeChoice.staple)) {
-      for (var mode in select_modes) {
-        selectors.add(':not(.${SelectModeChoice.scaffold.css_selector()}).${mode.css_selector()}');
-      }
-    }
-  } else {
-    // not origami, but origami and both scaffold and staples are selectable, just select anything
-    // with the correct mode. Otherwise use the more complex logic above to pick apart between
-    // parts marked as scaffold versus those that are not.
-    for (var mode in select_modes_not_scaffold_or_staple) {
-      selectors.add('.${mode.css_selector()}');
-    }
-  }
-  List<Element> selectable_elts = querySelectorAll(selectors.join(', '));
-  return selectable_elts;
-}
-
-bool bboxes_intersect_generalized(
-    Rect elt_bbox, Rect select_box_bbox, bool overlap(num l1, num h1, num l2, num h2)) {
-  num elt_x2 = elt_bbox.x + elt_bbox.width;
-  num select_box_x2 = select_box_bbox.x + select_box_bbox.width;
-  num elt_y2 = elt_bbox.y + elt_bbox.height;
-  num select_box_y2 = select_box_bbox.y + select_box_bbox.height;
-  return overlap(elt_bbox.x, elt_x2, select_box_bbox.x, select_box_x2) &&
-      overlap(elt_bbox.y, elt_y2, select_box_bbox.y, select_box_y2);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // unit testing utilities
 
 /// Returns the default state of the app.
